+++ pkg/discovery/memcache/resolver.go
@@ -92,10 +92,10 @@ func (s *memcachedAutoDiscovery) parseConfig(reader *bufio.Reader) (*clusterConf
 	}
 
 	if len(configVersion)+len(nodes) != configSize {
-		return nil, fmt.Errorf("expected %d in config payload, but got %d instead.", configSize, len(configVersion)+len(nodes))
+		return nil, fmt.Errorf("expected %d in config payload, but got %d instead", configSize, len(configVersion)+len(nodes))
 	}
 
-	for _, host := range strings.Split(strings.TrimSpace(nodes), " ") {
+	for host := range strings.SplitSeq(strings.TrimSpace(nodes), " ") {
 		dnsIpPort := strings.Split(host, "|")
 		if len(dnsIpPort) != 3 {
 			return nil, fmt.Errorf("node not in expected format: %s", dnsIpPort)